Flytter en del av logikken fra SendRegisterPapirSykmeldingOppgaveRest.kt ut i egen fil
